+++ modules/client/index.js
@@ -16,18 +16,37 @@ var mergeProperties = require('../utils/mergeProperties');
 ].forEach(function (method) {
   var methodName = method.toLowerCase();
 
-  mach[methodName] = function (app, options, callback) {
-    if (typeof app === 'string' || app instanceof Location) {
-      app = createProxy(app);
-    } else if (typeof app !== 'function') {
-      throw new Error('mach.' + methodName + ' needs an app');
+  mach[methodName] = function (app, options, modifier) {
+    if (typeof app !== 'function') {
+      modifier = options;
+
+      if (typeof app === 'string') { // get(url, modifier)
+        options = { url: app };
+      } else if (app instanceof Location) { // get(location, modifier)
+        options = { location: app };
+      } else { // get(options, modifier)
+        options = mergeProperties({}, app);
+      }
+
+      if (options.url || options.location) {
+        app = createProxy(options.url || options.location);
+      } else {
+        throw new Error('mach.' + methodName + ' needs an app');
+      }
+    } else if (typeof options === 'string') { // get(app, url, modifier)
+      options = { url: options };
+    } else if (options instanceof Location) { // get(app, location, modifier)
+      options = { location: options };
+    } else if (typeof options !== 'object') { // get(app, modifier)
+      modifier = options;
+      options = {};
+    } else { // get(app, options, modifier)
+      options = mergeProperties({}, options);
     }
 
-    // Don't mutate options argument.
-    options = options ? mergeProperties({}, options) : {};
     options.method = method;
 
-    return mach.call(app, options, callback);
+    return mach.call(app, options, modifier);
   };
 });