+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### HEAD
 
+  * Fixed client argument coercion, see #62
   * Added setters for Message#mediaType and Message#charset
   * Location#concat uses the protocol, auth, and host of its argument, if present
   * Location#protocol and Location#hostname default to null